Updater works better need to fix input modals
